+++ src/modules/server/server.service.ts
@@ -688,7 +688,7 @@ export class ServerService {
     return new Promise((res) => {
       let result = false
 
-      const child = spawn(process.execPath, ['-v'])
+      const child = spawn(process.execPath, ['-v'], { shell: true })
 
       child.stdout.once('data', (data) => {
         result = data.toString().trim() !== process.version